Gen_AI_2nd/requirements.txt
langchain==0.0.312
google-cloud-aiplatform
tabulate
pandas
db-dtypes
streamlit